Modify shortcut tests
